Updated plots so they don't clog the bandwidth/memory

+++ arcreactor/analysis.py
@@ -19,14 +19,14 @@ class Analyzer:
     def stream_names(self):
         return {'Reactor': ['plot']}
 
-    def get_plot(self, name, simulation_state, start_plotting):
-        if(start_plotting):
-            return self.plot_reactors(simulation_state)
-        else:
+    def get_plot(self, name, simulation_state, start_plotting, restart_plots):
+        if(restart_plots):
             self.xdata = []
             self.reactors = []
             self.reactor_number = 0
             self.start_time = simulation_state.time
+        if(start_plotting):
+            return self.plot_reactors(simulation_state)
         return None
 
     def plot_reactors(self,simulation_state):
@@ -51,26 +51,32 @@ class Analyzer:
             self.start_time = simulation_state.time
             self.xdata = []
             self.reactors = []
-        fig, axes = plt.subplots(len(simulation_state.kinetics), 1,  sharex = True, sharey = True, squeeze=False) #overlay all the plots
-        labels = ['A', 'B', 'C', 'D']
-        colors = ['red', 'blue', 'green', 'purple']
-        linestyles = ['-', ':', '-', ':']
+
         x = (simulation_state.time - self.start_time)*0.04      #display time in seconds(considering ~25fps)
         self.xdata.append(x)
-        i = 0
-        for i,ax in enumerate(axes[:,0]):
+        for i in range(len(simulation_state.kinetics)):
             y = simulation_state.kinetics[i].mole_fraction
             if(len(self.reactors) == i):#add one if we're out of space
                 self.reactors.append([[] for j in range(len(y))])
-            #self.reactors[i] = [[] for j in range(len(y))]
             for j in range(len(y)):#reactors stores the concentrations of each species of reactant in each reactor over time
                 self.reactors[i][j].append(y[j])#append this time step's molefrac to the respective reactor/species
-            for j in range(len(y)):
-                #print(self.xdata, ydata[:,j])
-                ax.plot(self.xdata, self.reactors[i][j], color = colors[j], label = labels[j], ls=linestyles[j])
-            plt.legend()
 
-        with io.BytesIO() as output:
-            fig.savefig(output, format='jpg')
-            plt.clf()
-            return output.getvalue()
+        if(simulation_state.time % 25 == 0):
+            fig, axes = plt.subplots(len(simulation_state.kinetics), 1,  sharex = True, sharey = True, squeeze=False) #overlay all the plots
+            fig.text(0.5, 0.04, 'Time Elapsed (s)', ha='center', va='center')
+            fig.text(0.01, 0.5, 'Concentration (mol/L)', ha='center', va='center', rotation='vertical')
+            labels = ['A', 'B', 'C', 'D']
+            colors = ['red', 'blue', 'green', 'purple']
+            linestyles = ['-', ':', '-', ':']
+            i = 0
+            for i,ax in enumerate(axes[:,0]):
+                #self.reactors[i] = [[] for j in range(len(y))]
+                for j in range(len(y)):
+                    #print(self.xdata, ydata[:,j])
+                    ax.plot(self.xdata, self.reactors[i][j], color = colors[j], label = labels[j], ls=linestyles[j])
+
+            plt.legend(loc='upper right')
+            with io.BytesIO() as output:
+                fig.savefig(output, format='jpg')
+                plt.clf()
+                return output.getvalue()
